runtime_requirements.txt
astor==0.8.1
attrdict==2.0.1
Babel==2.10.3
bce-python-sdk==0.8.74
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
contourpy==1.0.5
cssselect==1.1.0
cssutils==2.6.0
cycler==0.11.0
Cython==0.29.32
decorator==5.1.1
dill==0.3.5.1
et-xmlfile==1.1.0
Flask==2.2.2
Flask-Babel==2.0.0
Flask-MySQLdb==1.0.1
fonttools==4.37.3
future==0.18.2
idna==3.4
imageio==2.22.0
imgaug==0.4.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
jarowinkler==1.2.2
Jinja2==3.1.2
kiwisolver==1.4.4
lmdb==1.3.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.0
multiprocess==0.70.13
mysql-connector-python==8.0.30
mysqlclient==2.1.1
networkx==2.8.6
numpy==1.23.3
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
paddle-bfloat==0.1.7
paddleocr==2.6.0.1
paddlepaddle==2.3.2
pandas==1.5.0
Pillow==9.2.0
premailer==3.10.0
protobuf==3.20.0
pyclipper==1.3.0.post3
pycryptodome==3.15.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.4.1
rapidfuzz==2.10.1
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.1
Shapely==1.8.4
six==1.16.0
tifffile==2022.8.12
tqdm==4.64.1
urllib3==1.26.12
visualdl==2.4.1
Werkzeug==2.2.2
zipp==3.8.1
